requirements.txt
django>=4.2.0
djangorestframework>=3.14.0
torch>=2.0.0
torchvision>=0.15.0
Pillow>=9.0.0
requests>=2.31.0 
numpy>=1.24.0
pandas>=2.0.0
corsheaders>=2.0.0